// projects/yousuck/src/app/pages/avatar/avatar.types.ts
export interface Beard {
  key: string;
  name: string;
}
export interface Ear {
  key: string;
  name: string;
}
export interface Earring {
  key: string;
  name: string;
}
export interface Eyebrows {
  key: string;
  name: string;
}

export interface Eyes {
  key: string;
  name: string;
}
export interface Glasses {
  key: string;
  name: string;
}
export interface Hair {
  key: string;
  name: string;
}
export interface Mouth {
  key: string;
  name: string;
}
export interface Nose {
  key: string;
  name: string;
}
export interface Shirt {
  key: string;
  name: string;
}